backend/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic

# LangChain
langchain
langchain-google-genai
langchain-community

# Vector DB
faiss-cpu

# Data
pandas